Extract perm-go into separate repository
We were running into issues when depending on the vendored GRPC here in
this directory. This was because the generated go code from the perm
protobufs was using a vendored GRPC library, and any other client code
that depended on the public interfaces could not satisfy the vendored
library dependency.

Extracting the generated go code and having each client vendor the
interface definitions means they also vendor the GRPC library and
gogoproto, and we get around our problem.

Signed-off-by: Bruce Ricard <bricard@pivotal.io>