Consolidate DB connection logic within sqlx
This moves the core database connection logic out of SQLFlag.Connect and
into sqlx.Connect, such that the latter is now responsible for both
opening the connection and pinging the database to make sure that the
connection is viable.

A couple notes:
- SQLFlag.Connect is still temporarily around. Ideally, this can now be
removed completely, but that can be done independent of this commit.
- Similarly, DB.Conn still needs to be exported so that the migrator
tests can work as they are today (i.e., with sqlmock). We can - and
probaly should - change those to use real database connections, so that
we are confirming that the migrator works with all supported database
types. When we do this, we can also stop exposing DB.Conn.