Use unique indices in assignment tables
Historically, we hashed the relevant columns instead of using regular
unique indices because the length of the concatenated columns
potentially exceeded the max length allowed by mysql-based databases
that did not have `innodb_large_prefix` set to `true` and
`innodb_default_row_format` set to dynamic.

The defaults for mysql 5.7 and mariadb 10.2 are set correctly, but not
for 10.1. We no longer need to support mariadb 10.1, nor the upgrade
from 10.1. Therefore, we can go with simple unique indices :)

[Finishes #157527204]

Signed-off-by: Jeff Pak <jeffrey.pak@emc.com>